Made the overAll app to a standalone MediaPlayer component
