+++ client/src/MediaPlayer.css
@@ -1,0 +1,18 @@
+#playerTop{
+  padding-right: 5px;
+}
+#playerTop button{
+  border: none;
+  background-color: black;
+  color: whitesmoke;
+  margin: 2px;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
